% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dedup.R
\name{dedup}
\alias{dedup}
\title{Deduplicate a function on its first argument}
\usage{
dedup(f)
}
\arguments{
\item{f}{Function to deduplicate.}
}
\value{
Deduplicated function
}
\description{
Deduplicate a function on its first argument
}
\examples{

x <- sample(LETTERS, 10)
x

large_x <- sample(rep(x, 100))
length(large_x)

long_func <- function(x) for(i in x) {Sys.sleep(0.001)}

system.time({y <- long_func(large_x)})
system.time({y2 <- dedup(long_func)(large_x)})
all(y == y2)
}
